Set info section colour (also from Neodark)
Wombat matches the foreground colour to the current mode's background
which, as nice as it looks, is a bit distracting for the info section.
Bubblegum sets it to grey, but uses the default background (transparent
in my case) so that's out too.

Neodark has the best of both worlds with a muted foreground on a
background that's slightly brighter than the status line.