Set paste-mode insert colours
This uses the orange from Bubblegum as the mode background, but since
we're using that for the modified statusline foreground, we go with
Wombat's bright green for the actual filename section to avoid any
confusion. It looks a bit odd with orange, green AND blue, but since
paste mode is used so rarely it probably doesn't matter at all.